security-tests/Pytest/requirements.txt
requests
PyJWT
pytest
allure-pytest==2.13.5
openapi3
PyYAML
